requirements.txt
Django~=3.2.6
django-widget-tweaks~=1.4.8
django-allauth~=0.41.0